Move FatalImplementationException to ch.psc.exceptions
